✨ refactor(logs): renomear método e melhorar lógica de parada
Alterações realizadas para aprimorar a funcionalidade de parada do bot e a estrutura do código.

Files Changed:
- *api/namespaces/logs.py*: renomeado o método `on_stop_signal` para `on_stopbot` e ajustada a lógica para emitir um sinal de parada.
- *crawjud/abstract/_head.py*: adicionado um getter e setter para a propriedade `stop_bot`.
- *crawjud/bot.py*: removido código desnecessário relacionado ao controle de tarefas.
- *crawjud/bots/pje/capa.py*: implementada verificação de `stop_bot` em loops de execução.
- *crawjud/wrapper.py*: reorganizada a lógica de emissão do evento `stopbot`.
